+++ libraries/apk-editor/src/apksigner/io/ZipOutput.java
@@ -15,8 +15,6 @@
  */
 package apksigner.io;
 
-import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.List;
@@ -59,14 +57,14 @@ public class ZipOutput
     {
         CentralEnd centralEnd = new CentralEnd();
         
-        centralEnd.centralStartOffset = (int)getFilePointer();
+        centralEnd.centralStartOffset = getFilePointer();
         centralEnd.numCentralEntries = centralEnd.totalCentralEntries = (short)entriesWritten.size();
         
         for (ZioEntry entry : entriesWritten) {
             entry.write( this);
         }
         
-        centralEnd.centralDirectorySize = (int)(getFilePointer() - centralEnd.centralStartOffset);
+        centralEnd.centralDirectorySize = getFilePointer() - centralEnd.centralStartOffset;
         centralEnd.fileComment = "";
         
         centralEnd.write( this);
